refactor: move useRootData to hooks
