+++ client/paperflow-web-v1/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './App';
-import { StoreProvider } from './stores/useRootData';
+import { StoreProvider } from './hooks/useRootData';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
